requirements.txt
# FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data Validation and Settings
pydantic>=2.0.0
python-dotenv>=1.0.0

# HTTP Client
aiohttp>=3.8.0

# Google Cloud and ADK
google-adk>=0.1.0

# Utilities
pytz>=2023.3